backend/requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
ast2json==0.3
asttokens==2.0.8
astunparse==1.6.3
async-generator==1.10
async-timeout==4.0.2
attrs==22.1.0
awscli==1.27.35
backcall==0.2.0
bcrypt==4.0.1
beautifulsoup4==4.11.1
bleach==5.0.1
blinker==1.6.2
botocore==1.29.35
bs4==0.0.1
cachetools==5.3.0
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.4
contourpy==1.0.7
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
dj-database-url==0.5.0
dnspython==2.3.0
docutils==0.16
einops==0.6.0
entrypoints==0.4
exceptiongroup==1.1.1
executing==1.0.0
fastapi==0.95.0
fastjsonschema==2.16.1
filelock==3.8.0
Flask==2.3.2
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
Flask-PyMongo==2.3.0
flatbuffers==23.3.3
fonttools==4.39.3
frozenlist==1.3.3
gast==0.4.0
google-auth==2.17.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.54.0
h11==0.14.0
h5py==3.8.0
huggingface-hub==0.11.1
idna==3.4
importlib-metadata==4.12.0
importlib-resources==5.12.0
ipykernel==6.15.3
ipython==8.5.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
itsdangerous==2.1.2
jax==0.4.8
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonschema==4.16.0
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyter_client==7.3.5
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
keras==2.12.0
kiwisolver==1.4.4
libclang==16.0.0
linkedin-api==1.1.0
lxml==4.9.1
Markdown==3.4.3
MarkupSafe==2.1.1
matplotlib==3.7.1
matplotlib-inline==0.1.6
mistune==2.0.4
ml-dtypes==0.1.0
multidict==6.0.4
nbclient==0.6.8
nbconvert==7.0.0
nbformat==5.5.0
nest-asyncio==1.5.5
nltk==3.8.1
notebook==6.4.12
numpy==1.23.5
oauthlib==3.2.2
openai==0.27.0
opt-einsum==3.3.0
outcome==1.2.0
packaging==21.3
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==2.5.4
prometheus-client==0.14.1
prompt-toolkit==3.0.31
protobuf==4.22.3
psutil==5.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
puremagic==1.11
pyasn1==0.4.8
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.7
Pygments==2.13.0
PyJWT==2.7.0
pymongo==4.3.3
pyparsing==3.0.9
pypng==0.0.21
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2021.3
PyYAML==5.4.1
pyzmq==24.0.0
qtconsole==5.3.2
QtPy==2.2.0
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.7.2
s3transfer==0.6.0
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
selenium==4.8.3
Send2Trash==1.8.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
stack-data==0.5.0
starlette==0.26.1
tensorboard==2.12.2
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.2.0
terminado==0.15.0
textblob==0.17.1
threadpoolctl==3.1.0
tinycss2==1.1.1
tokenizers==0.13.2
torch==1.13.0
tornado==6.2
tqdm==4.64.1
traitlets==5.4.0
transformers==4.25.1
trio==0.22.0
trio-websocket==0.10.2
typing_extensions==4.4.0
urllib3==1.26.13
virtualenv==20.17.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.3.4
whitenoise==5.3.0
widgetsnbextension==4.0.3
wincertstore==0.2
wrapt==1.14.1
wsproto==1.2.0
yarl==1.8.2
zipp==3.8.1
